fix: reset 훅 분리 및 리셋 안되는 버그 수정
