+++ frontend/src/pages/EditChecklistPage.tsx
@@ -16,8 +16,8 @@ import { ROUTE_PATH } from '@/constants/routePath';
 import { DEFAULT_CHECKLIST_TAB_PAGE } from '@/constants/system';
 import useGetChecklistDetailQuery from '@/hooks/query/useGetChecklistDetailQuery';
 import useModal from '@/hooks/useModal';
+import useResetChecklist from '@/hooks/useResetChecklist';
 import useRoomInfoNonValidated from '@/hooks/useRoomInfoNonValidated';
-import roomInfoNonValidatedStore from '@/store/roomInfoNonValidatedStore';
 import roomInfoStore from '@/store/roomInfoStore';
 import useChecklistStore from '@/store/useChecklistStore';
 import useSelectedOptionStore from '@/store/useSelectedOptionStore';
@@ -32,26 +32,20 @@ const EditChecklistPage = () => {
   const { checklistId } = useParams() as RouteParams;
   const { data: checklist, isSuccess } = useGetChecklistDetailQuery(checklistId);
 
-  const checklistQuestionActions = useChecklistStore(state => state.actions);
+  const { resetChecklist } = useResetChecklist();
 
   const { searchSubwayStationsByAddress, set } = useRoomInfoNonValidated();
   const roomInfoActions = useStore(roomInfoStore, state => state.actions);
-  const roomInfoUnvalidatedActions = useStore(roomInfoNonValidatedStore, state => state.actions);
+  const checklistQuestionActions = useChecklistStore(state => state.actions);
+  const selectedOptionActions = useSelectedOptionStore(state => state.actions);
 
   // 한줄평 모달
   const { isModalOpen: isSubmitModalOpen, openModal: summaryModalOpen, closeModal: summaryModalClose } = useModal();
-
   // 메모 모달
   const { isModalOpen: isMemoModalOpen, openModal: memoModalOpen, closeModal: memoModalClose } = useModal();
 
-  // TODO: action 분리 필요
-  const selectedOptionActions = useSelectedOptionStore(state => state.actions);
-
   const resetAndGoDetailPage = () => {
-    roomInfoActions.reset();
-    roomInfoUnvalidatedActions.resetAll();
-    checklistQuestionActions.reset();
-    selectedOptionActions.reset();
+    resetChecklist();
     navigate(ROUTE_PATH.checklistOne(Number(checklistId)));
   };
 
@@ -62,7 +56,6 @@ const EditChecklistPage = () => {
       roomInfoActions.setRawValues(checklist.room);
       set('address', checklist.room.address!);
       set('buildingName', checklist.room.buildingName!);
-      //TODO: 가까운 지하철은 나중에 api 수정되면 저장
 
       loadExternalScriptWithCallback('kakaoMap', () => searchSubwayStationsByAddress(checklist.room.address!));
 
@@ -95,7 +88,6 @@ const EditChecklistPage = () => {
       )}
 
       {/* 한줄평 모달*/}
-
       <SubmitModalWithSummary
         isModalOpen={isSubmitModalOpen}
         onConfirm={resetAndGoDetailPage}
